Separate SpellCastMessageOverlay into a separate feature from the spell cast vignette
